requirements.txt
# 核心依賴
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.20.0
transformers>=4.30.0
accelerate>=0.20.0

# Stable Diffusion 相關
xformers>=0.0.20
invisible-watermark>=0.1.5
safetensors>=0.3.1

# ControlNet
controlnet-aux>=0.0.6
opencv-python>=4.8.0
mediapipe>=0.10.0

# 圖像處理
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# 動畫與GIF處理
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# 配置文件處理
PyYAML>=6.0
omegaconf>=2.3.0

# 進度條與日誌
tqdm>=4.65.0
rich>=13.0.0

# Web UI (可選)
gradio>=3.40.0

# 批量處理
joblib>=1.3.0

# 其他工具
requests>=2.31.0
beautifulsoup4>=4.12.0 